test: split agent lifecycle test cases
