+++ test/unit/agy_agent_hook_event_normalizer_test_cases.dart
@@ -1,0 +1,46 @@
+part of 'agent_hook_event_normalizer_test.dart';
+
+void _registerAgyAgentHookEventNormalizerTests() {
+  test('keeps AGY working until Stop reports fully idle', () {
+    expect(
+      normalizeAgentHookEvent(
+        _event(
+          agentType: AgentType.agy,
+          hookEventName: 'Stop',
+          payload: const <String, Object?>{'fullyIdle': false},
+        ),
+      )?.state,
+      AgentStatusState.working,
+    );
+    expect(
+      normalizeAgentHookEvent(
+        _event(
+          agentType: AgentType.agy,
+          hookEventName: 'Stop',
+          payload: const <String, Object?>{'fully_idle': false},
+        ),
+      )?.state,
+      AgentStatusState.working,
+    );
+    expect(
+      normalizeAgentHookEvent(
+        _event(
+          agentType: AgentType.agy,
+          hookEventName: 'Stop',
+          payload: const <String, Object?>{'fullyIdle': true},
+        ),
+      )?.state,
+      AgentStatusState.done,
+    );
+    expect(
+      normalizeAgentHookEvent(
+        _event(
+          agentType: AgentType.agy,
+          hookEventName: 'Stop',
+          payload: const <String, Object?>{},
+        ),
+      )?.state,
+      AgentStatusState.done,
+    );
+  });
+}
